extend report length

+++ class/hid/usbd_hid.c
@@ -76,7 +76,7 @@ __WEAK uint8_t usbh_hid_get_protocol(uint8_t intf)
     return 0;
 }
 
-__WEAK void usbh_hid_set_report(uint8_t intf, uint8_t report_id, uint8_t report_type, uint8_t *report, uint8_t report_len)
+__WEAK void usbh_hid_set_report(uint8_t intf, uint8_t report_id, uint8_t report_type, uint8_t *report, uint32_t report_len)
 {
 }
