+++ class/hid/usbd_hid.h
@@ -23,7 +23,7 @@ void usbd_hid_report_descriptor_register(uint8_t intf_num, const uint8_t *desc, 
 void usbh_hid_get_report(uint8_t intf, uint8_t report_id, uint8_t report_type, uint8_t **data, uint32_t *len);
 uint8_t usbh_hid_get_idle(uint8_t intf, uint8_t report_id);
 uint8_t usbh_hid_get_protocol(uint8_t intf);
-void usbh_hid_set_report(uint8_t intf, uint8_t report_id, uint8_t report_type, uint8_t *report, uint8_t report_len);
+void usbh_hid_set_report(uint8_t intf, uint8_t report_id, uint8_t report_type, uint8_t *report, uint32_t report_len);
 void usbh_hid_set_idle(uint8_t intf, uint8_t report_id, uint8_t duration);
 void usbh_hid_set_protocol(uint8_t intf, uint8_t protocol);
